skills/gerald-sniper/requirements.txt
# Жёсткие зависимости (только это, ничего лишнего):
pandas>=2.0
numpy>=1.24
scipy>=1.11
ta>=0.11
aiohttp>=3.9
aiosqlite>=0.19
aiogram>=3.4
mplfinance>=0.12
matplotlib>=3.7
pyyaml>=6.0
loguru>=0.7
